+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.2.0]
+
+### Additions
+
+- Added multi-track type
+
+### Changes
+
+- `start` and `end` track class parameters have moved and now have default None
+  values (**potentially breaking change**)
+
 ## [0.1.1]
 
 ### Additions
@@ -10,4 +21,5 @@ All notable changes to this project will be documented in this file.
 - Humorous cows
 - Backoff/Retries on requests that seem to occasionally fail
 
+[0.2.0]: https://github.com/James-Ansley/pipe-utils/compare/v0.1.2...v0.2.0
 [0.1.1]: https://github.com/James-Ansley/pipe-utils/compare/v0.0.1-alpha.1...v0.1.1